+++ Client.Linq/Internal/QueryExpressionTreeVisitor.cs
@@ -396,6 +396,7 @@ namespace InfluxDB.Client.Linq.Internal
                     .ToList();
 
                 foreach (var index in indexes)
+                {
                     // ()
                     if (parts.Count > index + 1 && parts[index + 1] is RightParenthesis)
                     {
@@ -405,6 +406,16 @@ namespace InfluxDB.Client.Linq.Internal
                         NormalizeExpressions(parts);
                         return;
                     }
+
+                    // (
+                    if (parts.Count == 1 && parts[index] is LeftParenthesis)
+                    {
+                        parts.RemoveAt(index);
+
+                        NormalizeExpressions(parts);
+                        return;
+                    }
+                }
 
                 // (( ))
                 if (parts.Count >= 4 && parts[0] is LeftParenthesis && parts[1] is LeftParenthesis &&